+++ vm-rust/src/player/geometry.rs
@@ -1,20 +1,20 @@
 use crate::console_warn;
 
 pub struct IntRect {
-  pub left: i16,
-  pub top: i16,
-  pub right: i16,
-  pub bottom: i16,
+  pub left: i32,
+  pub top: i32,
+  pub right: i32,
+  pub bottom: i32,
 }
 
-pub type IntRectTuple = (i16, i16, i16, i16);
+pub type IntRectTuple = (i32, i32, i32, i32);
 
 impl IntRect {
-  pub const fn from(l: i16, t: i16, r: i16, b: i16) -> IntRect {
+  pub const fn from(l: i32, t: i32, r: i32, b: i32) -> IntRect {
     return IntRect { left: l, top: t, right: r, bottom: b };
   }
 
-  pub const fn from_size(x: i16, y: i16, width: i16, height: i16) -> IntRect {
+  pub const fn from_size(x: i32, y: i32, width: i32, height: i32) -> IntRect {
     return IntRect::from(x, y, x + width, y + height);
   }
 
@@ -23,10 +23,10 @@ impl IntRect {
   }
 
   pub fn from_quad(
-    top_left: (i16, i16),
-    top_right: (i16, i16),
-    bottom_right: (i16, i16),
-    bottom_left: (i16, i16),
+    top_left: (i32, i32),
+    top_right: (i32, i32),
+    bottom_right: (i32, i32),
+    bottom_left: (i32, i32),
   ) -> IntRect {
     if top_left.1 != top_right.1 || top_right.0 != bottom_right.0 || bottom_right.1 != bottom_left.1 || bottom_left.0 != top_left.0 {
       console_warn!("INVALID IntRect::from_quad(({}, {}), ({}, {}), ({}, {}), ({}, {}))", top_left.0, top_left.1, top_right.0, top_right.1, bottom_right.0, bottom_right.1, bottom_left.0, bottom_left.1);
@@ -41,11 +41,11 @@ impl IntRect {
     };
   }
 
-  pub const fn width(&self) -> i16 {
+  pub const fn width(&self) -> i32 {
     return self.right - self.left;
   }
 
-  pub const fn height(&self) -> i16 {
+  pub const fn height(&self) -> i32 {
     return self.bottom - self.top;
   }
